+++ adminer/sqlite.php
@@ -1,9 +1,8 @@
 <?php
 function adminer_object() {
-	include_once "../plugins/login-password-less.php";
 	return new Adminer\Plugins(array(
 		// TODO: inline the result of password_hash() so that the password is not visible in source codes
-		new AdminerLoginPasswordLess(password_hash("YOUR_PASSWORD_HERE", PASSWORD_DEFAULT)),
+		new Adminer\Password(password_hash("YOUR_PASSWORD_HERE", PASSWORD_DEFAULT)),
 	));
 }