+++ editor/sqlite.php
@@ -2,8 +2,6 @@
 // see ../plugins/editor-setup.php for an easier solution
 
 function adminer_object() {
-	include_once "../plugins/login-password-less.php";
-
 	class AdminerCustomization extends Adminer\Plugins {
 		function loginFormField($name, $heading, $value) {
 			return parent::loginFormField($name, $heading, str_replace("value='server'", "value='sqlite'", $value));
@@ -15,7 +13,7 @@ function adminer_object() {
 
 	return new AdminerCustomization(array(
 		// TODO: inline the result of password_hash() so that the password is not visible in source codes
-		new AdminerLoginPasswordLess(password_hash("YOUR_PASSWORD_HERE", PASSWORD_DEFAULT)),
+		new Adminer\Password(password_hash("YOUR_PASSWORD_HERE", PASSWORD_DEFAULT)),
 	));
 }